custom eval input
also fix tokenizer encoding bug

+++ eval2.py
@@ -2,8 +2,7 @@ import multiprocessing
 import threading
 import time
 import uuid
-from collections import deque
-from queue import Empty, Full, ShutDown
+import queue
 
 import dearpygui.dearpygui as dpg
 import numpy as np
@@ -55,10 +54,24 @@ class LLNCAProcessManager:
         for process in self.processes:
             process.stop()
 
+    def add_input(self, user_input: str):
+        for process in self.processes:
+            try:
+                process.inputs.put(user_input, timeout=0.1)
+            except (queue.Full, queue.ShutDown):
+                continue
+
     def poll(self):
+        prev_user_input = ""
         prev_process_ids = []
         while True:
             time.sleep(LLNCAProcessManager.POLL_RATE)
+
+            user_input = dpg.get_value("user_input_field")
+            if user_input != prev_user_input:
+                self.add_input(user_input)
+            prev_user_input = user_input
+
             with self.processes_lock:
                 if not dpg.is_item_hovered("llnca_process_list"):
                     self.processes = [
@@ -105,16 +118,17 @@ class LLNCAProcessManager:
 
 class LLNCAProcess:
     POLL_RATE = 0.1
-    FRAME_RATE = 0.25
+    FRAME_RATE = 0.1
 
     def __init__(self, model_path: str):
         self.model_path = model_path
         self.id = str(uuid.uuid4())
+        self.inputs = multiprocessing.Queue()
         self.frames = multiprocessing.Queue()
         self.process_stop_event = multiprocessing.Event()
         self.process = multiprocessing.Process(
             target=self.process_run,
-            args=(self.model_path, self.process_stop_event, self.frames),
+            args=(self.model_path, self.process_stop_event, self.inputs, self.frames),
             daemon=True,
         )
         self.thread_stop_event = threading.Event()
@@ -141,8 +155,17 @@ class LLNCAProcess:
         print(f"\033[2mstopping {self.model_path}\033[0m")
         self.process_stop_event.set()
         self.thread_stop_event.set()
+
+        self.process.join(timeout=0.5)
+        if self.process.is_alive():
+            self.process.terminate()
+            self.process.join()
+
+        self.inputs.close()
+        self.frames.close()
+        self.inputs.join_thread()
         self.frames.cancel_join_thread()
-        self.process.terminate()
+
         dpg.delete_item(self.id + "_window")
 
     def thread_run(self, stop_event, frames: multiprocessing.Queue):
@@ -150,7 +173,7 @@ class LLNCAProcess:
         while not stop_event.is_set():
             try:
                 frame = frames.get(timeout=LLNCAProcess.POLL_RATE)
-            except Empty:
+            except queue.Empty:
                 continue
             dpg.set_value(self.id + "_x_str", frame["x_str"])
             dpg.set_value(self.id + "_y_str", frame["y_str"])
@@ -160,28 +183,41 @@ class LLNCAProcess:
             prev_frame = frame["frame"]
 
     @staticmethod
-    def process_run(model_path: str, stop_event, frames: multiprocessing.Queue):
+    def process_run(
+        model_path: str,
+        stop_event,
+        inputs: multiprocessing.Queue,
+        frames: multiprocessing.Queue,
+    ):
         checkpoint = torch.load(
             model_path, weights_only=False, map_location=torch.device("cpu")
         )
         llnca = LLNCA(checkpoint=checkpoint)
-        x_str, y_str = "", ""
-        prev_x, prev_y = "", ""
-        for x, y, xs, frame in llnca.eval():
-            if stop_event.is_set():
+
+        x_str = ""
+        xs = None
+        while not stop_event.is_set():
+            try:
+                x_str = inputs.get(timeout=LLNCAProcess.POLL_RATE)
+                xs = llnca.eval_start([x_str])
+            except queue.ShutDown:
                 break
-            x_str = llnca.tokenizer.decode(x) if x != prev_x else x_str
-            y_str = llnca.tokenizer.decode(y) if y != prev_y else y_str
+            except queue.Empty:
+                pass
+
+            if not x_str or xs is None:
+                continue
+
+            xs, y_pred = llnca.eval_step(xs)
+            frame = llnca.tokenizer.decode(y_pred[0])
+
             try:
                 frames.put(
-                    {"x_str": x_str, "y_str": y_str, "frame": frame},
+                    {"x_str": x_str, "y_str": x_str, "frame": frame},
                     timeout=LLNCAProcess.POLL_RATE,
                 )
-            except (Full, ShutDown):
+            except queue.ShutDown:
                 break
-            prev_x, prev_y = x, y
-
-        frames.close()
 
     def get_state_desc(self):
         stopping = self.process_stop_event.is_set()
@@ -207,12 +243,12 @@ if __name__ == "__main__":
 
     dpg.set_exit_callback(exit_callback)
 
-    # with dpg.font_registry():  # type: ignore
-    #     myfont = dpg.add_font(
-    #         "/usr/share/fonts/truetype/JetBrainsMono/JetBrainsMonoNerdFont-Medium.ttf",
-    #         16,
-    #     )
-    # dpg.bind_font(myfont)
+    with dpg.font_registry():  # type: ignore
+        myfont = dpg.add_font(
+            "/usr/share/fonts/truetype/JetBrainsMono/JetBrainsMonoNerdFont-Medium.ttf",
+            16,
+        )
+    dpg.bind_font(myfont)
 
     with dpg.file_dialog(
         tag="model_path_picker",
@@ -226,9 +262,27 @@ if __name__ == "__main__":
         dpg.add_file_extension(".tar")
         dpg.add_file_extension(".*")
 
-    with dpg.window(tag="main_window", label="llnca eval suite"):  # type: ignore
+    with dpg.window(
+        tag="main_window",
+        label="llnca eval suite",
+        width=300,
+        height=200,
+        pos=[10, 10],
+    ):  # type: ignore
         dpg.add_button(
             label="load model", callback=lambda: dpg.show_item("model_path_picker")
+        )
+
+    with dpg.window(
+        tag="user_input",
+        label="input",
+        width=300,
+        height=200,
+        pos=[10, 220],
+    ):  # type: ignore
+        dpg.add_input_text(
+            tag="user_input_field",
+            hint="type smth...",
         )
 
     with dpg.window(  # noqa: SIM117
